+++ CHANGELOG.md
@@ -3,7 +3,10 @@
 All notable changes to `NovaPoshta` will be documented in this file
 
 ### 2022-11-07
+- [Add] Добавление метода `createApiKey` в `CommonGeneral`
+- [Add] Добавление метода `deleteApiKey` в `CommonGeneral`
 - [Add] Добавление метода `saveAddTerm` в `AdditionalService`
+- [Add] Добавление метода `saveChangeCash` в `AdditionalService`
 - [Add] Добавление метода `getDocumentPrice` в `InternetDocument`
 - [Fix] Убрал 3 и 4 параметр в методе `getDocumentDeliveryDate` в `InternetDocument` (Ломает обратку)